backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.2.0
dj-rest-auth==5.0.2
django-allauth==0.58.2
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-spectacular==0.27.0
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2
stripe==8.11.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.39.1

# Production dependencies
django-redis==5.4.0
python-decouple==3.8

# New feature dependencies
twilio==8.10.0
pywebpush==1.14.0
qrcode==7.4.2
geopy==2.4.1
pytz==2023.3

# Development dependencies (not needed in production)
django-debug-toolbar==4.2.0
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.4
black==23.12.0
flake8==6.1.0